Moved app_info to own package
