+++ app/AndroidManifest.xml
@@ -33,7 +33,7 @@
 
 
         <activity
-            android:name="de.ironjan.mensaupb.activities.About_"
+            android:name="de.ironjan.mensaupb.app_info.About_"
             android:label="@string/app_name"
             android:parentActivityName="de.ironjan.mensaupb.menus_ui.Menus_">
             <meta-data
