+++ app/src/main/java/de/ironjan/mensaupb/app_info/AboutFragment.java
@@ -1,4 +1,4 @@
-package de.ironjan.mensaupb.fragments;
+package de.ironjan.mensaupb.app_info;
 
 import android.content.Intent;
 import android.support.v4.app.Fragment;
