+++ app/src/main/res/layout/fragment_about.xml
@@ -1,6 +1,6 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="de.ironjan.mensaupb.AboutFragment"
+    tools:context="de.ironjan.mensaupb.app_info.AboutFragment_"
 
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
